Moved simulation code into own module.
