+++ msprime/__init__.py
@@ -29,6 +29,7 @@ from _msprime import REVERSE  # NOQA
 from msprime.environment import __version__  # NOQA
 from msprime.formats import *  # NOQA
 from msprime.trees import *  # NOQA
+from msprime.simulations import *  # NOQA
 from msprime.stats import *  # NOQA
 from msprime.exceptions import *  # NOQA
 from msprime.tables import *  # NOQA
